refactored flascard component
