+++ src/app/app-routing.module.ts
@@ -4,7 +4,7 @@ import { AuthenticationGuard } from './guard/authentication.guard';
 import { PeriodicTableComponent } from './periodic-table/periodic-table.component';
 import { UserComponent } from './user/user.component';
 import { QuizComponent } from './quiz/quiz.component';
-import { FlashcardComponent } from './components/flashcard/flashcard.component';
+import { FlashcardComponent } from './flashcard/flashcard.component';
 import { AuthorizationGuard } from './guard/authorization.guard';
 import { ProfileComponent } from './profile/profile.component';
 import { EditProfileComponent } from './edit-profile/edit-profile.component';
