--- src/app/components/flashcard/flashcard.component.html
@@ -1,50 +1,0 @@
-<!-- create flashcard -->
-<div class="container" style="margin-top: 100px;">
-  <div class="d-flex justify-content-center h-50">
-    <div class="user_card">
-      <div class="d-flex justify-content-center">
-        <div style="margin-top: 10px;margin-bottom:-90px;">
-          <h3>Create Flash Card</h3>
-        </div>
-      </div>
-      <div class="d-flex justify-content-center form_container">
-        <form #loginForm="ngForm" (ngSubmit)="createFlashcard(loginForm.value)">
-          <div class="input-group mb-3">
-            <input type="text" class="form-control" name="question" placeholder="Enter a question" ngModel
-              #questionInput="ngModel" required>
-          </div>
-          <span class="help-block" style="color:red;" *ngIf="questionInput.invalid && questionInput.touched">Please
-            enter a question</span>
-          <div class="input-group mb-2">
-            <input type="text" class="form-control" name="answer" placeholder="Enter the corret answer" ngModel
-              #answerInput="ngModel" required>
-          </div>
-          <span class="help-block" style="color:red;" *ngIf="answerInput.invalid && answerInput.touched">Please
-            enter a answer.</span>
-          <div class="d-flex justify-content-center mt-3 login_container">
-            <button type="submit" [disabled]="loginForm.invalid" name="button" class="btn login_btn">
-              <span>Submit</span>
-            </button>
-          </div>
-        </form>
-      </div>
-      <div class="mt-4">
-        <div class="d-flex justify-content-center links">
-          <a routerLink="/flashcard/all" class="ml-2" style="color: #2C3E50;">See All Flash Card(s)</a>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-
-<!-- list -->
-<div class="flashcard-container">
-  <div class="app-flash-card" *ngFor="let displayCard of flashcards; let i = index">
-    <div class="card-grid">
-      <div [ngClass]="flipped ? 'card flip' : 'card'" (click)="flipCard($event)">
-        <div #frontEl class="front">{{displayCard.question}}</div>
-        <div #backEl class="back">{{displayCard.answer}}</div>
-      </div>
-    </div>
-  </div>
-</div>
