+++ src/app/flashcard/flashcard.component.ts
@@ -1,5 +1,5 @@
 import { HttpErrorResponse, HttpResponse } from '@angular/common/http';
-import { Component, OnInit, Input, Output, EventEmitter, ViewChild, AfterViewInit, HostListener, ElementRef } from '@angular/core';
+import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
 import { Router } from '@angular/router';
 import { AuthenticationService } from 'src/app/service/authentication.service';
 import { FlashCardService } from 'src/app/service/flashcard.service';
@@ -11,16 +11,8 @@ import { FlashCardService } from 'src/app/service/flashcard.service';
 })
 export class FlashcardComponent implements OnInit {
   flashcards: any[] = [];
-
-  @ViewChild("frontEl")
-  frontEl!: ElementRef;
-
-  @ViewChild("backEl")
-  backEl!: ElementRef;
-
   flashcard: any = {};
-
-  flipped = false;
+  allCards: boolean = false;
 
   constructor(private router: Router, private service: FlashCardService, private authenticationService: AuthenticationService) { }
 
@@ -31,33 +23,28 @@ export class FlashcardComponent implements OnInit {
       this.getAllFlashcardsByUserId(this.authenticationService.getUserFromLocalCache().userId);
   }
 
-  flipCard(event: any) {
-    this.flipped = !this.flipped;
+  flipCard(id: number) {
+    if(!this.flashcards[id]['flipped']) {
+      document.getElementById(String(id)).classList.add("flip");
+      this.flashcards[id]['flipped'] = true;
+    } else {
+      document.getElementById(String(id)).classList.remove("flip");
+      this.flashcards[id]['flipped'] = false;
+    }
   }
 
   public createFlashcard(item: any): void {
     let userId = this.authenticationService.getUserFromLocalCache().userId;
     item[ "userId" ] = userId;
-    // console.log(item);
     this.service.createFlashcard(item).subscribe({
-      next: (response: HttpResponse<any>) => {       
+      next: (response: HttpResponse<any>) => {
+        this.getAllFlashcardsByUserId(this.authenticationService.getUserFromLocalCache().userId);
       },
       error: (errorResponse: HttpErrorResponse) => {
         console.error(errorResponse);
       }
     });
   }
-
-  // public getAllFlashcard(): void {
-  //   this.service.getAllFlashcard().subscribe(
-  //     (response: any) => {
-  //       this.flashcards = response;
-  //     },
-  //     (errorResponse: HttpErrorResponse) => {
-  //       console.error(errorResponse);
-  //     }
-  //   );
-  // }
 
   public getAllFlashcardsByUserId(userId: string): void {
     this.service.getFlashcardsByUserId(userId).subscribe({
@@ -68,6 +55,10 @@ export class FlashcardComponent implements OnInit {
         console.error(errorResponse);
       }
     });
+  }
+
+  displayCards() {
+    this.allCards = !this.allCards;
   }
 
 }
